Update routes incrementally
Instead of clearing the routes and adding them all back afterwards, we
only clear and init the routes once, at the start, then update them
incrementally:

- Delete the routes to hosts no longer available.
- Use "replace" instead of "add", wich allows to modify the route, if it
  exists, or else to create it.

The "replace" mecanism (almost atomic), combined with not clearing the
routes on each update, allows to keep the route in the table as long as
it exists, and prevents it from flickering, which in turn, prevents "no
route to host" errors when updating the routes at high frequency (e.g.
1Hz).

Next will be the tc commands, as I suspect dropping the qdisc instead of
replacing them also leads to a lot of packet losses.