WIP: do not flush tc ruleset, update it instead
This allows to make sure no packet is dropped while updating the link
properties, but it has the undesired effect that replacing a qdisc can
be blocked for an undefined amount of time.

By the way, use the newer "ether" match of u32 filter to simplify the
filter rules.

Commment out the client health code, as we do not really care about it
for now.