create one macvlan iface for each peer
This allows to make the TC rules a lot simpler, as we can get rid of the
classful qdisc, and thus, simply use a "change" instruction to change
the parameters of the netem qdisc. In the end, this allows to keep the
qdisc in place, and get rid of unwanted packet loss due to qdisc
removal.

By the way, we remove the TBF qdisc and only use the native netem rate
limiting capabilities. This allows to simplify the setup even further.