Stop relying on /tmp/gw file supposedly created by the container img
These "clear_" script are called only once at the start of the
topologiser, so there is no problem getting the infos once from iproute2
instead of getting them from a file supposedly created by another
process of the docker image.

By the way, retrieve the network prefix size instead of assuming that it
will always be a /16, which is sometimes incorrect, especially when
creating new networks often.