Allow running multiple instances of meshsim in parallel
Meshsim now can be run up to 10 times in parallel by using a network
ientifier between 0 and 9 (inclusive). A docker network is now created
on the fly for each instance of meshsim, and all the names/ports are
namespaced using this network identifier.

The instances are completely separated and run in their own sparate
virtual network.